feat: move the calibration function to a script
